Add other icon, rename, integrate
