+++ src/components/icons/VerifiedCheck.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
-import Svg, {Circle,Path} from 'react-native-svg'
+import Svg, {Circle, Path} from 'react-native-svg'
 
 import {type Props, useCommonSVGProps} from '#/components/icons/common'
 
-export const VerificationCheck = React.forwardRef<Svg, Props>(function LogoImpl(
+export const VerifiedCheck = React.forwardRef<Svg, Props>(function LogoImpl(
   props,
   ref,
 ) {
@@ -23,7 +23,7 @@ export const VerificationCheck = React.forwardRef<Svg, Props>(function LogoImpl(
         fill="#fff"
         fillRule="evenodd"
         clipRule="evenodd"
-        d="M18.203 7.51a1.5 1.5 0 0 1 0 2.122l-6.857 6.857a1.5 1.5 0 0 1-2.121 0l-3.429-3.428a1.5 1.5 0 1 1 2.122-2.122l2.368 2.368 5.796-5.796a1.5 1.5 0 0 1 2.121 0Z"
+        d="M18.311 7.421a1.437 1.437 0 0 1 0 2.033l-6.571 6.571a1.437 1.437 0 0 1-2.033 0L6.42 12.74a1.438 1.438 0 0 1 2.033-2.033l2.27 2.269 5.554-5.555a1.437 1.437 0 0 1 2.033 0Z"
       />
     </Svg>
   )
