+++ src/components/verification/VerificationCreatePrompt.tsx
@@ -5,7 +5,7 @@ import {useLingui} from '@lingui/react'
 
 import {atoms as a} from '#/alf'
 import {type DialogControlProps} from '#/components/Dialog'
-import {VerificationCheck} from '#/components/icons/VerificationCheck'
+import {VerifiedCheck} from '#/components/icons/VerifiedCheck'
 import * as Prompt from '#/components/Prompt'
 
 export function VerificationCreatePrompt({
@@ -21,7 +21,7 @@ export function VerificationCreatePrompt({
   return (
     <Prompt.Outer control={control}>
       <View style={[a.flex_row, a.align_center, a.gap_sm, a.pb_sm]}>
-        <VerificationCheck width={18} />
+        <VerifiedCheck width={18} />
         <Prompt.TitleText style={[a.pb_0]}>
           {_(msg`Verify ${userName}`)}
         </Prompt.TitleText>
